requirements.txt
streamlit
python-dotenv
pypdf
langchain
langchain-community
langchain-openai
openai
chromadb
